Fix compiler warnings for cmake build
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@166677 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: a26167fa45c1b043c6c6cea63925410ceb8d1fa6

+++ Online/Online/Gaucho/src/ObjSerializer.cpp
@@ -129,7 +129,7 @@ void *ObjSerializer::SerializeObj(std::vector<std::string> &nams,void * &ptr,int
     else
     {
       DimBuffBase hh;
-      hh.type = H_ILLEGAL;
+      hh.type = (unsigned int)H_ILLEGAL;
       memcpy (pp,&hh,sizeof(hh));
       pp = AddPtr(pp,sizeof(DimBuffBase));
     }
